backend/requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib
python-multipart
httpx
typing
sqlalchemy
typing
pydantic